openrisc/mm: move protection fault detection to do_page_fault
Instead of clearing the 'vector' parameter to do_page_fault()
to indicate that a fault was caused by a non-protection error,
make do_page_fault figure this out on its own.

This is done by peeking into the TLBs to see if the page that
caused the fault is not cached in any of the TLB entries
(i.e. a TLB miss).
This works, because only a page fault generated at the same time
as a TLB hit can be a protection fault, and vice versa.

The motivation for this is that, the 'vector' parameter is
cleared by the TLB miss exceeption vectors, but in order to
support hardware page walks (which are supported by the
architecture) one can not rely on the non-protective
page fault coming through the TLB miss vectors.

Signed-off-by: Stefan Kristiansson <stefan.kristiansson@saunalahti.fi>